created end2end test with creation+del of dataset + test matrix
